Move code to .NET Fx
